+++ qemu/tcg/tcg.c
@@ -244,8 +244,8 @@ void tcg_context_init(TCGContext *s)
     s->code_buf = code_gen_prologue;
     s->code_ptr = s->code_buf;
     tcg_target_qemu_prologue(s);
-    flush_icache_range((unsigned long)s->code_buf, 
-                       (unsigned long)s->code_ptr);
+    flush_icache_range((uintptr_t)s->code_buf, 
+                       (uintptr_t)s->code_ptr);
 }
 
 void tcg_set_frame(TCGContext *s, int reg,
@@ -1968,7 +1968,7 @@ static inline int tcg_gen_code_common(TCGContext *s, uint8_t *gen_code_buf,
             break;
         case INDEX_op_set_label:
             tcg_reg_alloc_bb_end(s, s->reserved_regs);
-            tcg_out_label(s, args[0], (long)s->code_ptr);
+            tcg_out_label(s, args[0], (intptr_t)s->code_ptr);
             break;
         case INDEX_op_call:
             dead_iargs = s->op_dead_iargs[op_index];
@@ -2017,8 +2017,8 @@ int tcg_gen_code(TCGContext *s, uint8_t *gen_code_buf)
     tcg_gen_code_common(s, gen_code_buf, -1);
 
     /* flush instruction cache */
-    flush_icache_range((unsigned long)gen_code_buf, 
-                       (unsigned long)s->code_ptr);
+    flush_icache_range((uintptr_t)gen_code_buf, 
+                       (uintptr_t)s->code_ptr);
     return s->code_ptr -  gen_code_buf;
 }
